+++ src/main/java/team9499/commitbody/global/authorization/dto/response/TokenUserInfoResponse.java
@@ -1,8 +1,9 @@
-package team9499.commitbody.global.authorization.dto;
+package team9499.commitbody.global.authorization.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import team9499.commitbody.global.authorization.dto.TokenInfoDto;
 
 @Data
 @AllArgsConstructor
